requirements.txt
et_xmlfile==2.0.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
setuptools==80.9.0
six==1.17.0
tzdata==2025.2
wheel==0.45.1
